+++ Core/Frameworks/BaikalAdmin/Controller/Login.php
@@ -1,4 +1,5 @@
 <?php
+
 #################################################################
 #  Copyright notice
 #
@@ -24,11 +25,9 @@
 #  This copyright notice MUST APPEAR in all copies of the script!
 #################################################################
 
-
 namespace BaikalAdmin\Controller;
 
 class Login extends \Flake\Core\Controller {
-
     function execute() {
     }
 
@@ -37,9 +36,15 @@ class Login extends \Flake\Core\Controller {
         $sSubmittedFlagName = "auth";
         $sMessage = "";
 
+        $sLogin = htmlspecialchars(\Flake\Util\Tools::POST("login"));
+
         if (self::isSubmitted() && !\BaikalAdmin\Core\Auth::isAuthenticated()) {
-            // Log failed accesses, for further processing by other tools (fail2ban)
-            error_log('user not authorized: Baikal GUI');
+            // Log failed accesses, matching the default fail2ban nginx/apache auth rules
+            if (isset($_SERVER['SERVER_SOFTWARE']) && preg_match('/nginx/i', $_SERVER['SERVER_SOFTWARE'])) {
+                error_log('user "' . $sLogin . '" was not found in "Baikal GUI"', 4);
+            } else {
+                error_log('user "' . $sLogin . '" authentication failure for "Baikal GUI"', 4);
+            }
             $sMessage = \Formal\Core\Message::error(
                 "The login/password you provided is invalid. Please retry.",
                 "Authentication error"
@@ -52,7 +57,6 @@ class Login extends \Flake\Core\Controller {
             );
         }
 
-        $sLogin = htmlspecialchars(\Flake\Util\Tools::POST("login"));
         $sPassword = htmlspecialchars(\Flake\Util\Tools::POST("password"));
 
         if (trim($sLogin) === "") {
@@ -75,6 +79,7 @@ class Login extends \Flake\Core\Controller {
 
     protected static function justLoggedOut() {
         $aParams = $GLOBALS["ROUTER"]::getURLParams();
+
         return (!empty($aParams) && $aParams[0] === "loggedout");
     }
 }